Update pathToCSV and added 'MS Program Complete'
The `pathToCSV` was updated so that this is an option that can be passed to the nm_loadTR() function.

Also, 'MS Program Complete' was added to the type of successful runs that are processed. For lift-and-drag testing, the model-support (MS) program complete is a successful event that results in normal shutdown.