Pull AoA info into TRdata struct.
The data was already there in SRV400, but I'm pulling this into the main struct for cases where the number of unique MS_Pt values is greater than 1. I'm assuming that for "MS Not Used" cases there would only be one value for MS_Pt. I may need to revisit this later.